[spine-csharp] fixed compilation errors initially not reported by Mono CS

+++ spine-csharp/src/TransformConstraint.cs
@@ -77,7 +77,7 @@ namespace Spine {
 
 				if (rotateMix > 0) {
 					float a = bone.a, b = bone.b, c = bone.c, d = bone.d;
-					float r = Math.Atan2(tc, ta) - Math.Atan2(c, a) + data.offsetRotation * MathUtils.degRad;
+					float r = (float)Math.Atan2(tc, ta) - (float)Math.Atan2(c, a) + data.offsetRotation * MathUtils.degRad;
 					if (r > MathUtils.PI)
 						r -= MathUtils.PI2;
 					else if (r < -MathUtils.PI) r += MathUtils.PI2;
